+++ public/index.html
@@ -31,6 +31,6 @@
     <!-- Количество фильмов -->
   </section>
 </footer>
-<script type="text/javascript" src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
