+++ src/view/extra-list.js
@@ -1,0 +1,10 @@
+export const createExtraListTemplate = (title) => {
+  return (`
+    <section class="films-list--extra">
+      <h2 class="films-list__title">${title}</h2>
+
+      <div class="films-list__container">
+      </div>
+    </section>
+  `);
+};
