+++ src/view/footer-statistics-value.js
@@ -1,0 +1,5 @@
+export const createFooterStatisticsValueTemplate = () => {
+  return (`
+    <p>130 291 movies inside</p>`
+  );
+};
