+++ src/view/show-more-btn.js
@@ -1,0 +1,5 @@
+export const createShowMoreBtn = () => {
+  return (`
+    <button class="films-list__show-more">Show more</button>
+  `);
+};
